Store a copy of gdCurrentArticle as ArticleView's data member
This is the fourth in a series of commits that get rid of
evaluateJavaScript() calls. The rationale is given in the first such
commit "Don't call gdCheckArticlesNumber() from C++ code".

Don't call gdArticleView.onJsActiveArticleChanged() from embedded
JavaScript code, which runs before gdArticleView.onPageJsReady(). This
call lead to a current dictionary list item update in
MainWindow::activeArticleChanged(). This update was useless because the
dictionary list is completely reset at the end of
ArticleView::loadFinished() - in MainWindow::pageLoaded(). Instead pass
gdCurrentArticle as a second parameter to onPageJsReady().