Print JavaScript console messages as warnings
JavaScript errors appear in GoldenDict's output now instead of being
ignored. Makes debugging JavaScript much easier.

A beneficial side effect of the added QWebView::setPage() call is that
the initial blank page is no longer present in QWebHistory. Remove the
now obsolete workaround in ArticleView::canGoBack() to allow going back
to the Welcome! page in the main window and to the first loaded page in
the scan popup.

Adjust the javascript href code on a "(picture)" page to allow going
back on image click if the previous page was the first one in a tab or
in the scan popup. Replace `history.go(-1)` with the more obvious
equivalent `history.back()`. Now that the blank page is absent from web
history, history.back() can be called unconditionally, because the call
does nothing if there is no previous page.

By the way, the removed `history.length>2` check was unreliable before
this commit: when the user opened a gdpicture link in a new tab (e.g.
via a middle button click), then loaded another page in this new tab,
then went back to the "(picture)" page and clicked on the image, the new
tab's blank page was unintentionally navigated to.