Don't call gdCheckArticlesNumber() from C++ code
The following harmless warning at GoldenDict start is now eliminated:
JS: ReferenceError: Can't find variable: gdCheckArticlesNumber (at undefined:1)

QWebEnginePage::runJavaScript() is Qt WebEngine's equivalent of
QWebFrame::evaluateJavaScript(). This equivalent looks similar but is
far less useful in practice: being asynchronous, it provides no ordering
guarantees. See also:
https://wiki.qt.io/QtWebEngine/ScriptsAndExtensions#Script_Injection_in_Qt_WebEngine

This is the first in a series of commits that get rid of
evaluateJavaScript() calls in order to make execution more predictable,
prevent timing-related issues and improve performance of the upcoming Qt
WebEngine version of GoldenDict.

Load and call the gdCheckArticlesNumber() JavaScript function in a new
script file with the `defer` attribute. A deferred script is executed
after the document has been parsed, but before the DOMContentLoaded
event is fired. So the JavaScript code in deferred.js is evaluated after
all blocking and directly-embedded JavaScipt code, but before the
QWebView::loadFinished() signal is emitted and
ArticleView::loadFinished() is invoked.

Use Immediately Invoked Function Expression as the function needs to be
called exactly once at the point of definition. An alternative is to
move the function body into the global namespace, but that would pollute
the global namespace with variables.