Store loadedPageHasSound as ArticleView's data member
This is the second in a series of commits that get rid of
evaluateJavaScript() calls. The rationale is given in the first such
commit "Don't call gdCheckArticlesNumber() from C++ code".

The JavaScript code is executed and the added function onPageJsReady()
is invoked before ArticleView::loadFinished(). So loadedPageHasSound is
set early enough - before the slots connected to
ArticleView::pageLoaded() call ArticleView::hasSound().

Make ArticleView's JavaScript interface private and add a friend
ArticleViewJsProxy. Enforcing and documenting that these functions are
invoked only from JavaScript (and potentially from other member
functions of ArticleView itself) facilitates understanding of the code.
ArticleViewJsProxy is a tiny class that only forwards messages from
JavaScript. So making it a friend doesn't harm encapsulation much.