Store a copy of gdArticleContents as ArticleView's data member
This is the sixth in a series of commits that get rid of
evaluateJavaScript() calls. The rationale is given in the first such
commit "Don't call gdCheckArticlesNumber() from C++ code".

Make gdArticleContents an array of strings rather than a long string
with space separators. This change shouldn't have much impact on
performance, but leads to more concise and less error-prone code.

Rename getArticlesList() to getArticleList() (English grammar) and make
it const.