Allow building against Qt WebEngine
Remove `#if QT_VERSION >= 0x040600` along the way as GoldenDict does not
support Qt versions older than 4.6 for several years now.

The Qt WebEngine version requires Qt 5.15.2 or later. Qt 5.15.2 was
released almost two years ago and is available in latest versions of
most major GNU/Linux distributions.

The Qt WebEngine version requires C++14 support from the compiler. This
is a rather conservative requirement for the standard version released
more than 7 years ago. Major compilers support C++14 for many years now.
Note that Qt 5.7 requires C++11 and Qt 6.0 requires C++17.

QWebEngineHistoryItem has no equivalent of QWebHistoryItem::userData()
API. I have tried to store the current article and the coordinates in a
container data member of ArticleView. The container had a designated
position and mirrored web history. This implementation was shared with
the Qt WebKit version. But it turned out that the mirroring of web
history is not an easy task. The stored entries and position quickly got
out of sync with QWebEngineHistory when the user requested several
translations in quick succession, because not every URL, that started to
load and appeared in ArticleView::handleUrlChanged(), ended up in web
history.

So the JavaScript History API is used in the Qt WebEngine version.
Unfortunately this API does not work correctly in the Qt WebKit version,
not even in the latest Qt 5.15.5 with KDE patches and QtWebKit 5.212.0
Alpha 4. Thus the two versions use different web history APIs.

This commit borrows some ideas from @xiaoyifang's unmerged porting pull
request #1447.