Print a warning when JavaScript sends invalid active article ID
This should never happen. If it does, the warning could be helpful.

In general, it is a good idea to validate data received from JavaScript,
because nothing prevents malicious code on a third-party website from
calling gdArticleView.onJsActiveArticleChanged('<garbage>') as long as
gdArticleView lives in the same world as the page's web contents
(QWebEngineScript::MainWorld).