selectCurrentArticle: don't pass current article to JavaScript
There is a copy of ArticleView::currentArticle on the JavaScript side:
gdCurrentArticle. Normally the values of these two variables are equal.
In the rare situations when they differ, gdCurrentArticle should take
precedence because it is more likely to be up-to-date.

Examples:
1. Load a page with no articles (e.g. a not-found page) and request
   translation of a phrase that is present in many active dictionaries.
   Press Ctrl+Shift+A before the page with articles is fully loaded.
   Before this commit the entire page was selected. Now the topmost
   article is selected.
2. Press Ctrl+Shift+A before the article from the dictionary, which was
   current on the previous page, is loaded. Before this commit a
   JavaScript Type error message was printed and nothing was selected.
   Now the topmost article is selected.

Another advantage of not passing the current article from C++ to
JavaScript applies only to the upcoming Qt WebEngine version: when the
user clicks on an article's text to make it current, then very quickly
presses Ctrl+Shift+A, the interprocess communication may not have
completed, and the previously current article could be selected. This
should not be a problem in the other direction, because messages from
C++ to JavaScript should arrive in the order they are sent.