WebEngine-specific: find and set page background color
This commit eliminates white background flashes when the article style's
background color is not white.

ArticleNetworkAccessManager::getResource() obtains a single string (not
counting contentType, which should not be abused to return the
background color). So getResource() cannot be used to retrieve both the
blank page itself and the background color. Add a special-purpose
function ArticleNetworkAccessManager::makeBlankPage() for this.

Limit the background-finding code to the Qt WebEngine version to be able
to use std::regex from C++11 (neither QRegExp nor QRegularExpression
support 8-bit strings). QWebPage does not support setting the background
color anyway. Though one could set QWebPage::palette instead if needed,
given std::regex availability.

Setting the syntax of article URL schemes to Host is no longer necessary
to prevent a white background flash => remove the comment. Keep the Host
syntax because the now-invisible flash may be caused by some
security-related reloading, which might waste the CPU time. The URLs
constructed in the code contain the host (localhost), so this is the
correct syntax.

Remove commented out body selector from built-in article-style.css file.
This was the only piece in the built-in style sheets that prevented
correct regex matching of the <body> background color.