Update Back and Forward buttons on fragment navigation
ArticleView::loadFinished() is not called for fragment navigation on the
same page. Back and Forward buttons are not enabled/disabled according
to web history state then.

ArticleView::handleUrlChanged() is called for all URL changes, including
fragment navigation on the same page. So canGoBackForwardChanged() could
be emitted from this one place instead of the current two. However,
QWebEngineHistory::canGo* return outdated results at this early stage of
the location change when a new phrase translation is requested.