WebEngine-specific: set persistent data and cache paths
Replace the default paths in the portable version and on GNU/Linux.

Move the applyWebSettings() call lower in the MainWindow constructor.
This allows to set up the web engine profile after the article network
access manager, which should be ready for ArticleUrlSchemeHandler; and
before editing the profile's settings, which will be necessary in the
future when setupWebEngineProfile() creates a custom profile.