WebEngine-specific: forbid navigating to initial blank page
The blank page is present in web history when no other page is
immediately loaded in its place:
  1) in the scan popup;
  2) in a new tab created via the New Tab action.

A "(picture)" page can be the first non-blank page in web history only
in a tab created by opening a gdpicture link in a new tab. In this case
the blank page is absent from the web history - just as in the Qt WebKit
version. So no adjustment to the javascript href code in
ArticleMaker::makePicturePage() is needed.

I have tried to set isBlankPagePresentInWebHistory in
ArticleView::onJsPageInitStarted() instead of
ArticleView::handleUrlChanged(). But if the next page is loaded fast
enough (Ctrl+T followed immediately by V without releasing Ctrl to
paste a phrase from clipboard), handleUrlChanged() is and
onJsPageInitStarted() is not called for the blank page, after which the
blank page is present in the web history and can be navigated back to.