ArticleViewJsProxy: Q_INVOKABLE => slot
Unlike a Q_INVOKABLE function, a slot cannot return a value. JavaScript
proxy's member functions should not return values, because JavaScript
code is executed in a separate process in the upcoming Qt WebEngine
version.

Another advantage is that when more member functions are added to
ArticleViewJsProxy, they won't have to be prefixed with Q_INVOKABLE.