Eliminate gdlookup://localhost/bullet.gif errors
The bullet.gif image referenced in the .mwiki style in article-style.css
is not available in GoldenDict repository or online. The closest image I
could find: https://commons.wikimedia.org/wiki/File:Bullet.png

Even had the image been available, the "url(bullet.gif)" reference to it
would not work. In ArticleNetworkAccessManager::getResource() neither
group nor phrase of this URL is valid, meaning that getResource()
returns a null pointer and ArticleNetworkAccessManager::createRequest()
forwards the request to QNetworkAccessManager::createRequest(), which
does not know about the "gdlookup" protocol.

So the simplest fix is to remove the references to this nonexistent
image from article-style.css.

This commit eliminates an error message each time a page containing
MediaWiki articles is loaded:
* in the Qt WebEngine version, one of the following warnings in
  GoldenDict's output:
  Error while requesting URL gdlookup://localhost/bullet.gif: Protocol "gdlookup" is unknown (ProtocolUnknownError)
  Error while requesting URL gdlookup://localhost/bullet.gif: Operation canceled (OperationCanceledError)
* in the Qt WebKit version, an error in Web Inspector:
  Failed to load resource: Protocol "gdlookup" is unknown - gdlookup://localhost/bullet.gif