WebEngine-specific: optimize finding page background color
Add a new custom CSS property --gd-page-background-color into built-in
CSS files and require this property to be specified in all user-defined
CSS files, except for print-only ones. Show a UI warning message at each
GoldenDict start if the variable is unspecified in one or more CSS files
or if the color value is invalid. The prominent and detailed warning
message should annoy most users into adding the CSS property to their
custom style sheets in short order.

Even though the current code reads CSS files line by line until the CSS
property specification is found, and the property is specified close to
the top of each CSS file, the overall size of a CSS file still
determines the search time. The time spent in QFile::open() depends on
the file size and dominates the search time for some reason. Neither
QIODevice::Text nor QIODevice::Unbuffered flags passed to QFile::open()
have a noticeable impact on this time.

The search for page background color was and remains slowest when the
Default display style is configured. That's because article-style.css is
the largest of the article CSS files, and all built-in styles specify
the page background color; so when another built-in style is configured,
article-style.css is never opened.

This commit reduces the longest search time on my GNU/Linux system from
about 15 ms to less than 0.4 ms.

I have tried an alternative method of finding the page background color:
1. Embed a script into articles right after appending CSS style sheets.
2. The script retrieves the page's background color with this code:
    window.getComputedStyle(document.documentElement).backgroundColor
3. The script sends the result to ArticleView, which forwards it to
   MainWindow.
4. MainWindow compares the received background color with a previously
   stored one. If the colors differ, the new color is stored and passed
   to each existing article view, as well as each article view created
   in the future.
5. Each ArticleView passes the updated background color to
   QWebEnginePage::setBackgroundColor().
Important advantages of this alternative method are:
1) it is accurate and reliable;
2) does not require changing user-defined article styles.
However, the alternative method also has significant drawbacks:
1. If the "Start to system tray" option is off (the default), a white
   background flash appears in the initial Welcome! page, because the
   page background color is determined and set after the page becomes
   visible.
2. Computing the document element's or the body's style in JavaScript
   takes more than 2 ms on average, which is 5 times slower than the
   worst-case scenario at this commit.
3. The CPU time waste is even worse if the background-color-computing
   script is embedded into each page. Embedding the script only in the
   initial Welcome! page and possibly into blank pages causes background
   flashes when a different Display style is selected or a user-defined
   article style is modified while GoldenDict is running.
In the end I decided against the alternative method, because its
drawbacks are permanent. Its advantages exist only until user-defined
styles are adjusted, and after that - only while the styles are being
tweaked, which shouldn't happen often.