Save Article: find and process custom resource links recursively
The next commit will move JavaScript code embedded in article_maker.cc
to a new file blocking.js and reference the js file in an article page's
HTML code instead of embedding the JavaScript code into HTML. The
extracted js file will contain custom (qrcx://) resource links. So when
a complete article is saved, GoldenDict has to look for and process such
links in the same way as it processes links in the main HTML file.

Handle the progress dialog cancellation properly to prevent leaks and
crashes.

Optimize path manipulation by ending absolute and relative destination
paths with a slash and caching them.