WebEngine-specific: port [un]setting web view cursor
[Un]setting QWebEngineView's cursor does not affect the actual mouse
pointer appearance. [Un]setting its widget child's cursor makes the
mouse pointer look the same as in the Qt WebKit version.

Don't fall back to hidden QWebEngineView's equivalents in the added
cursor member functions if childWidget is null. If QWebEngineView's
cursor is set to Qt::WaitCursor before childWidget becomes nonnull, and
childWidget's cursor is "unset" after childWidget becomes nonnull, the
QWebEngineView's cursor will remain WaitCursor indefinitely. A visible
effect would be: when an article is loaded, the cursor does not restore
the default Qt::ArrowCursor shape, Qt::WaitCursor is displayed until the
user vigorously moves the mouse pointer. The described issue occurs in
practice only if the assignment to childWidget is moved into a lambda
invoked by a queued connection in ChildAdded event. But even a
possibility of this issue surfacing after a future code change is enough
to avoid the fallback.