Display an indicator while a page is loading
The primary motivation for this new feature is to mitigate QTBUG-106580.
When a page is not fully loaded, loadFinished() is never invoked. So the
loading-state indicator remains visible until a next page replaces this
one and finishes loading. The user has a chance to notice the stuck
loading indicator, conclude that something went wrong and realize that
the page may be incomplete. And then, hopefully, the user would reload
the page, e.g. by going back and forward in web history.

The loading-state indicator's placement in the UI is dictated by the
following requirements:
1) always present in the main window and the scan popup UI, no matter
   how the user configures GoldenDict's UI;
2) does not waste valuable UI space;
3) does not disrupt any dictionary usage workflows.

I have implemented a proof-of-concept animated icon version, but decided
against it because of potentially significant CPU time waste. Plus the
animation could be distracting. Pages are loaded often enough that the
users should quickly learn to recognize the static loading-state icon.

I have implemented an alternative mitigation via overriding the
application cursor with a busy cursor instead of the current setting of
a wait cursor over the article web view. A cursor set via
QWidget::setCursor() is replaced with a regular cursor on mouse move. An
overridden application cursor stays until restored or overridden again.
However I rejected this approach, because it does not work if the user
prefers keyboard navigation and does not keep the mouse pointer over
GoldenDict windows.

The added icon loading.gif is distributed under the GPLv3 license and
can be downloaded at
https://github.com/shlinux/faenza-icon-theme/blob/master/emesene/themes/Faenza/loading.gif