WebEngine-specific: share SelectWordBySingleClick script
The selectWordBySingleClick preference value is common to all pages.
Instead of inserting and updating a script that defines the
gdSelectWordBySingleClick JavaScript variable in each article web page's
script collection, insert and update it in the shared web engine
profile's script collection.

Name the shared script "ProfilePreferences" to facilitate extending it
by injecting other shared preferences in the future.